2.1-flask/requirements.txt
flask
requests
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
Flask-Bcrypt
Flask-HTTPAuth
